pyproject.toml
[tool.poetry]
name = "morocco-delivery-mas"
version = "0.1.0"
description = "AI-powered multi-agent delivery system for Morocco"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = {extras = ["tools"], version = ">=0.28.0"}
langchain-openai = ">=0.0.5"
pydantic = ">=2.0.0"
requests = ">=2.31.0"
pyyaml = ">=6.0"
python-dateutil = ">=2.8.2"
python-dotenv = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
black = ">=23.0.0"
flake8 = ">=6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
